requirements.txt
asgiref==3.2.7
Django==3.0.5
django-nested-admin==3.3.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.1.2
PyJWT==1.7.1
python-monkey-business==1.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
